[scripts: build.sh] Expand the build.sh script to also support dfu bin
This commit is about expanding the build tool to support binary post
processing to make the binary compatible with the bootloader (and the
bootloader firmware update) requirements. It now uses the
build_info.yaml to treat each project differently. Also, it now supports
a build.sh all command to build all the needed projects in one go.

Signed-off-by: George Pappas <pappasgeorge12@gmail.com>