[bootloader/uart_driver] Adding uart watchdog to recover uart it receive
This commit is about adding uart watchdog. The TIM1 is used to clear the
uart periodically, when the uart is inactive. It might be the case that
some bytes are received (less than 256). If that happens, uart will
still wait for the rest of the bytes until it receives 256, in order to
parse them. This can create a permanent issue in uart communication.

After a period of inactivity, uart is reinitialized to expect again 256
bytes.

Signed-off-by: George Pappas <pappasgeorge12@gmail.com>